I guess this is general WM improvements now.
Completely reconfigure the window management system and window tracking.

There are now three major items tracked in the HashMap: `name`, `WINDOW`, and `WindowData`. `WindowData` is the renamed `NewWindow`.  This was done to help with #43 and #45.

Reorganizes the `new_window` function into a few discrete parts. Firstly, `new_window` which creates, tracks, and draws a new window, and then a window drawing function separate, and then a redrawing function salso separate. Should reduce coupling and make the whole system a little more flexible.